+++ nengo/builder/ensemble.py
@@ -6,7 +6,7 @@ import numpy as np
 import nengo.utils.numpy as npext
 from nengo.builder import Builder, Signal
 from nengo.builder.operator import Copy, DotInc, Reset
-from nengo.dists import Distribution
+from nengo.dists import Distribution, get_samples
 from nengo.ensemble import Ensemble
 from nengo.neurons import Direct
 from nengo.utils.builder import default_n_eval_points
@@ -57,12 +57,6 @@ class BuiltEnsemble(collections.namedtuple('BuiltEnsemble', built_attrs)):
                                    max_rates, scaled_encoders, gain, bias))
 
 
-def sample(dist, n, d=None, rng=None):
-    if isinstance(dist, Distribution):
-        return dist.sample(n, d=d, rng=rng).astype(np.float64)
-    return np.array(dist, dtype=np.float64)
-
-
 def gen_eval_points(ens, eval_points, rng, scale_eval_points=True):
     if isinstance(eval_points, Distribution):
         n_points = ens.n_eval_points
@@ -89,8 +83,8 @@ def get_activities(params, ens, eval_points):
 
 def get_gain_bias(ens, rng=np.random):
     if ens.gain is not None and ens.bias is not None:
-        gain = sample(ens.gain, ens.n_neurons, rng=rng)
-        bias = sample(ens.bias, ens.n_neurons, rng=rng)
+        gain = get_samples(ens.gain, ens.n_neurons, rng=rng)
+        bias = get_samples(ens.bias, ens.n_neurons, rng=rng)
         max_rates, intercepts = None, None  # TODO: determine from gain & bias
     elif ens.gain is not None or ens.bias is not None:
         # TODO: handle this instead of error
@@ -98,8 +92,8 @@ def get_gain_bias(ens, rng=np.random):
                                   "Solving for one given the other is not "
                                   "implemented yet." % ens)
     else:
-        max_rates = sample(ens.max_rates, ens.n_neurons, rng=rng)
-        intercepts = sample(ens.intercepts, ens.n_neurons, rng=rng)
+        max_rates = get_samples(ens.max_rates, ens.n_neurons, rng=rng)
+        intercepts = get_samples(ens.intercepts, ens.n_neurons, rng=rng)
         gain, bias = ens.neuron_type.gain_bias(max_rates, intercepts)
 
     return gain, bias, max_rates, intercepts
@@ -152,7 +146,8 @@ def build_ensemble(model, ens):
     if isinstance(ens.neuron_type, Direct):
         encoders = np.identity(ens.dimensions)
     elif isinstance(ens.encoders, Distribution):
-        encoders = sample(ens.encoders, ens.n_neurons, ens.dimensions, rng=rng)
+        encoders = get_samples(
+            ens.encoders, ens.n_neurons, ens.dimensions, rng=rng)
     else:
         encoders = npext.array(ens.encoders, min_dims=2, dtype=np.float64)
     encoders /= npext.norm(encoders, axis=1, keepdims=True)